+++ NEWS.md
@@ -1,3 +1,5 @@
+# RAWdataR 1.2.4
+
 # RAWdataR 1.2.3
 
 * move header to the front in RAW-ID.csv file